[bug] Beim Loeschen von Shopeintraegen werden Bestellungen (laufende/bearbeitete) nun korrekt beruecksichtigt.

+++ src/net/driftingsouls/ds2/server/modules/ErsteigernController.java
@@ -2138,6 +2138,13 @@ public class ErsteigernController extends TemplateGenerator
 
 			if( "löschen".equalsIgnoreCase(getString("operation")) )
 			{
+				if( entry.getAnzahlOffeneBestellungen() > 0 )
+				{
+					addError("Es gibt noch offene Bestellungen zu diesem Shopeintrag");
+					redirect("shop");
+					return;
+				}
+
 				db.delete(entry);
 
 				t.setVar("show.message", "Eintrag gelöscht");
